# swagger.yaml
openapi: 3.0.0
info:
  title: Event Management API
  version: 1.0.0
  description: API documentation for the Event Management System
servers:
  - url: http://localhost:3000/api
paths:
  /events:
    post:
      summary: Create a new event
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the event
                  example: Annual Meeting
                description:
                  type: string
                  description: The event description
                  example: Discussing the annual results.
                date:
                  type: string
                  format: date-time
                  description: The date of the event (YYYY-MM-DD)
                  example: '2024-12-01'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid data
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                        location:
                          type: string
    get:
      summary: Retrieve all events
      tags:
        - Events
      responses:
        '200':
          description: A list of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the event
        schema:
          type: integer
    get:
      summary: Retrieve a specific event by ID
      tags:
        - Events
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Event not found
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                        location:
                          type: string
    put:
      summary: Update an event by ID
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the event
                  example: Updated Meeting
                description:
                  type: string
                  description: The event description
                  example: Discussing the updated results.
                date:
                  type: string
                  format: date-time
                  description: The date of the event (YYYY-MM-DD)
                  example: '2024-12-05'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Event not found
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                        location:
                          type: string
    delete:
      summary: Delete an event by ID
      tags:
        - Events
      responses:
        '204':
          description: Event deleted successfully
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Event not found
  /events/{id}/tasks:
    get:
      summary: Retrieve all tasks for a specific event
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the event
          schema:
            type: integer
      responses:
        '200':
          description: A list of tasks for the event
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: The unique identifier for the task
                    eventId:
                      type: integer
                      description: The ID of the associated event
                    name:
                      type: string
                      description: The name of the task
                    description:
                      type: string
                      description: A description of the task
                    dueDate:
                      type: string
                      format: date-time
                      description: The due date of the task
                    status:
                      type: string
                      description: The status of the task
                    createdAt:
                      type: string
                      format: date-time
                      description: The creation timestamp
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Event not found
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                        location:
                          type: string
...
components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the event
        name:
          type: string
          description: The name of the event
        description:
          type: string
          description: A description of the event
        date:
          type: string
          format: date-time
          description: The date and time of the event
        createdAt:
          type: string
          format: date-time
          description: The creation timestamp
...